Extract shared fuso-core crate and reorganize repo structure
