+++ .github/workflows/publish.yml
@@ -12,8 +12,12 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
-      - name: Publish
-        working-directory: cli
+      - name: Publish fuso-core
         env:
           CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
-        run: cargo publish
+        run: cargo publish -p fuso-core
+
+      - name: Publish fuso
+        env:
+          CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
+        run: sleep 30 && cargo publish -p fuso
