+++ .gitignore
@@ -6,5 +6,4 @@ DerivedData/
 .DS_Store
 Fuso.app/
 thoughts/
-cli/target/
-linux/target/
+target/
