+++ Cargo.toml
@@ -1,0 +1,7 @@
+[workspace]
+members = [
+    "crates/fuso-core",
+    "cli",
+    "linux",
+]
+resolver = "2"
