+++ cli/Cargo.toml
@@ -11,9 +11,6 @@ keywords = ["timezone", "cli", "worldclock", "team"]
 categories = ["command-line-utilities"]
 
 [dependencies]
+fuso-core = { path = "../crates/fuso-core" }
 chrono = "0.4"
 chrono-tz = "0.10"
-dirs = "6"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-iana-time-zone = "0.1"
