+++ crates/fuso-core/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "fuso-core"
+version = "0.1.0"
+edition = "2021"
+description = "Shared config, timezone, and availability logic for Fuso"
+license = "MIT"
+
+[dependencies]
+chrono = "0.4"
+chrono-tz = "0.10"
+dirs = "6"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+iana-time-zone = "0.1"
