+++ macos/Sources/Info.plist
